+++ tests/test_wheelfile.py
@@ -161,3 +161,7 @@ def test_attributes(tmpdir_factory, wheel_path):
             info = zf.getinfo(filename)
             assert info.external_attr == (mode | 0o100000) << 16
             assert info.compress_type == ZIP_DEFLATED
+
+        info = zf.getinfo('test-1.0.dist-info/RECORD')
+        permissions = (info.external_attr >> 16) & 0o777
+        assert permissions == 0o664